Restructured repository package. Turned off Spring liquibase.
