+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/repository/jdbc/ChatRowMapper.java
@@ -1,7 +1,8 @@
-package ru.tinkoff.edu.java.scrapper.repository;
+package ru.tinkoff.edu.java.scrapper.repository.jdbc;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
+import ru.tinkoff.edu.java.scrapper.repository.ChatRecord;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
