+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/jdbc/JdbcChatService.java
@@ -3,10 +3,12 @@ package ru.tinkoff.edu.java.scrapper.service.jdbc;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-import ru.tinkoff.edu.java.scrapper.repository.JdbcScrapperRepository;
+import ru.tinkoff.edu.java.scrapper.repository.ChatRecord;
+import ru.tinkoff.edu.java.scrapper.repository.jdbc.JdbcScrapperRepository;
 
-//@Service
-@Component
+import java.util.List;
+
+@Service
 public class JdbcChatService {
     @Autowired
     private JdbcScrapperRepository repository;
@@ -17,5 +19,9 @@ public class JdbcChatService {
 
     public void remove(int chatId) {
         repository.deleteChat(chatId);
+    }
+
+    public List<ChatRecord> getAll() {
+        return repository.getAllChats();
     }
 }
