+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/jdbc/JdbcLinkService.java
@@ -1,9 +1,8 @@
 package ru.tinkoff.edu.java.scrapper.service.jdbc;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-import ru.tinkoff.edu.java.scrapper.repository.JdbcScrapperRepository;
+import ru.tinkoff.edu.java.scrapper.repository.jdbc.JdbcScrapperRepository;
 import ru.tinkoff.edu.java.scrapper.repository.LinkRecord;
 
 import java.util.List;
